requirements.txt
streamlit>=1.28.0
opencv-python-headless>=4.7.0
mediapipe>=0.10.0
numpy>=1.24.0
pillow>=9.0.0
